requirements.txt
scikit-learn
pandas
numpy
keras
streamlit
openai
python-dotenv
matplotlib
fpdf